requirements.txt
# Telegram Bot Framework
aiogram==3.13.1

# Database
asyncpg==0.29.0
sqlalchemy==2.0.36
alembic==1.14.0

# Environment variables
python-dotenv==1.0.1

# Timezone handling
pytz==2024.2

# Retry logic
tenacity==9.0.0

# Date/time utilities
python-dateutil==2.9.0

# Logging (built-in, but can add structured logging later)
# structlog==24.1.0  # Optional for future

# ============== TESTING ==============
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-mock==3.12.0
pytest-cov==4.1.0
faker==22.0.0
aiosqlite==0.19.0
greenlet==3.0.0
